Use float for coordinates

+++ Object.cpp
@@ -3,7 +3,7 @@
 
 void Star::Randomize(unsigned int windowWidth, unsigned int windowHeight)
 {
-    pos = {std::rand() % (int)windowWidth, std::rand() % (int)windowHeight};
+    pos = {(float)(std::rand() % (int)windowWidth), (float)(std::rand() % (int)windowHeight)};
     speed = 1 + std::rand() % 8;
     height = 1 + std::rand() % 3;
     if(speed <= 2) {
